requirements.txt
streamlit
torch
opencv-python
Pillow
ultralytics
